# Improve websocket masking, use long datatype for XOR operation.