--- a/codec-http/src/main/java/io/netty/handler/codec/http/websocketx/WebSocket08FrameEncoder.java
+++ b/codec-http/src/main/java/io/netty/handler/codec/http/websocketx/WebSocket08FrameEncoder.java
@@ -83,7 +83,7 @@ public class WebSocket08FrameEncoder extends MessageToMessageEncoder<WebSocketFr
     private static final byte OPCODE_PONG = 0xA;
 
     /**
-     * The size threshold for gathering writes. Non-Masked messages bigger than this size will be be sent fragmented as
+     * The size threshold for gathering writes. Non-Masked messages bigger than this size will be sent fragmented as
      * a header and a content ByteBuf whereas messages smaller than the size will be merged into a single buffer and
      * sent at once.<br>
      * Masked messages will always be sent at once.
@@ -139,8 +139,7 @@ public class WebSocket08FrameEncoder extends MessageToMessageEncoder<WebSocketFr
         b0 |= opcode % 128;
 
         if (opcode == OPCODE_PING && length > 125) {
-            throw new TooLongFrameException("invalid payload for PING (payload length must be <= 125, was "
-                    + length);
+            throw new TooLongFrameException("invalid payload for PING (payload length must be <= 125, was " + length);
         }
 
         boolean release = true;
@@ -148,10 +147,7 @@ public class WebSocket08FrameEncoder extends MessageToMessageEncoder<WebSocketFr
         try {
             int maskLength = maskPayload ? 4 : 0;
             if (length <= 125) {
-                int size = 2 + maskLength;
-                if (maskPayload || length <= GATHERING_WRITE_THRESHOLD) {
-                    size += length;
-                }
+                int size = 2 + maskLength + length;
                 buf = ctx.alloc().buffer(size);
                 buf.writeByte(b0);
                 byte b = (byte) (maskPayload ? 0x80 | (byte) length : (byte) length);
@@ -168,7 +164,7 @@ public class WebSocket08FrameEncoder extends MessageToMessageEncoder<WebSocketFr
                 buf.writeByte(length & 0xFF);
             } else {
                 int size = 10 + maskLength;
-                if (maskPayload || length <= GATHERING_WRITE_THRESHOLD) {
+                if (maskPayload) {
                     size += length;
                 }
                 buf = ctx.alloc().buffer(size);
@@ -186,7 +182,6 @@ public class WebSocket08FrameEncoder extends MessageToMessageEncoder<WebSocketFr
                 ByteOrder srcOrder = data.order();
                 ByteOrder dstOrder = buf.order();
 
-                int counter = 0;
                 int i = data.readerIndex();
                 int end = data.writerIndex();
 
@@ -194,25 +189,35 @@ public class WebSocket08FrameEncoder extends MessageToMessageEncoder<WebSocketFr
                     // Use the optimized path only when byte orders match
                     // Remark: & 0xFF is necessary because Java will do signed expansion from
                     // byte to int which we don't want.
-                    int intMask = ((mask[0] & 0xFF) << 24)
-                                | ((mask[1] & 0xFF) << 16)
-                                | ((mask[2] & 0xFF) << 8)
-                                | (mask[3] & 0xFF);
+                    long longMask = ((long) mask[0] & 0xff) << 56
+                                    | ((long) mask[1] & 0xff) << 48
+                                    | ((long) mask[2] & 0xff) << 40
+                                    | ((long) mask[3] & 0xff) << 32
+                                    | ((long) mask[0] & 0xff) << 24
+                                    | ((long) mask[1] & 0xff) << 16
+                                    | ((long) mask[2] & 0xff) << 8
+                                    | (long) mask[3] & 0xff;
 
                     // If the byte order of our buffers it little endian we have to bring our mask
                     // into the same format, because getInt() and writeInt() will use a reversed byte order
                     if (srcOrder == ByteOrder.LITTLE_ENDIAN) {
-                        intMask = Integer.reverseBytes(intMask);
+                        longMask = Long.reverseBytes(longMask);
                     }
 
-                    for (; i + 3 < end; i += 4) {
-                        int intData = data.getInt(i);
-                        buf.writeInt(intData ^ intMask);
+                    for (int lim = end - 7; i < lim; i += 8) {
+                        long longData = data.getLong(i);
+                        buf.writeLong(longData ^ longMask);
+                    }
+
+                    if (i < end - 3) {
+                        buf.writeInt(data.getInt(i) ^ (int) (longMask >> 32));
+                        i += 4;
                     }
                 }
+                int maskOffset = 0;
                 for (; i < end; i++) {
                     byte byteData = data.getByte(i);
-                    buf.writeByte(byteData ^ mask[counter++ % 4]);
+                    buf.writeByte(byteData ^ mask[maskOffset++ & 3]);
                 }
                 out.add(buf);
             } else {