--- a/microbench/src/main/java/io/netty/microbench/websocket/package-info.java
+++ b/microbench/src/main/java/io/netty/microbench/websocket/package-info.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2022 The Netty Project
+ *
+ * The Netty Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+/**
+ * Benchmarks for {@link io.netty.handler.codec.http.websocketx.WebSocket08FrameDecoder}
+ * and {@link io.netty.handler.codec.http.websocketx.WebSocket08FrameEncoder}.
+ */
+package io.netty.microbench.websocket;