Motivation:

The performance of a websocket depends a lot on frame masking and unmasking, we can improve this by using a long data type for the XOR operation. 

Modification:

Use long datatype for XOR operation. This reduces the number of reads and writes required when masking a data buffer.
Provide benchmarks to check result.

Result:

Websocket performance is better.