@amizurov 2022-07-27
Encoder benchmarks result:

New encoder:
```
Benchmark                                                                              (contentLength)  (masking)  (pooledAllocator)   Mode  Cnt        Score        Error   Units
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                                 0       true               true  thrpt   20  6053405.457 ± 230633.076   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                  0       true               true  thrpt   20      131.933 ±      5.020  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                             0       true               true  thrpt   20       24.004 ±      0.012    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                         0       true               true  thrpt   20      131.061 ±     12.300  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                    0       true               true  thrpt   20       23.851 ±      2.090    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                     0       true               true  thrpt   20        0.001 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm                0       true               true  thrpt   20       ≈ 10⁻⁴                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                       0       true               true  thrpt   20       60.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                        0       true               true  thrpt   20       60.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                                 4       true               true  thrpt   20  5681735.799 ± 209720.013   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                  4       true               true  thrpt   20      123.843 ±      4.565  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                             4       true               true  thrpt   20       24.005 ±      0.013    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                         4       true               true  thrpt   20      124.517 ±     18.555  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                    4       true               true  thrpt   20       24.080 ±      3.158    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                     4       true               true  thrpt   20        0.001 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm                4       true               true  thrpt   20       ≈ 10⁻⁴                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                       4       true               true  thrpt   20       57.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                        4       true               true  thrpt   20       52.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                                 8       true               true  thrpt   20  5732408.241 ± 408951.058   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                  8       true               true  thrpt   20      124.964 ±      8.932  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                             8       true               true  thrpt   20       24.008 ±      0.024    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                         8       true               true  thrpt   20      126.696 ±     16.963  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                    8       true               true  thrpt   20       24.423 ±      3.347    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                     8       true               true  thrpt   20        0.001 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm                8       true               true  thrpt   20       ≈ 10⁻⁴                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                       8       true               true  thrpt   20       58.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                        8       true               true  thrpt   20       55.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                                32       true               true  thrpt   20  5035009.568 ± 335131.257   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                 32       true               true  thrpt   20      109.728 ±      7.298  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                            32       true               true  thrpt   20       24.001 ±      0.001    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                        32       true               true  thrpt   20      107.033 ±     19.361  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                   32       true               true  thrpt   20       23.378 ±      3.703    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                    32       true               true  thrpt   20        0.001 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm               32       true               true  thrpt   20       ≈ 10⁻⁴                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                      32       true               true  thrpt   20       49.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                       32       true               true  thrpt   20       42.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                               100       true               true  thrpt   20  5016865.289 ± 437914.956   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                100       true               true  thrpt   20      109.360 ±      9.545  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                           100       true               true  thrpt   20       24.005 ±      0.016    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                       100       true               true  thrpt   20      109.230 ±     19.464  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                  100       true               true  thrpt   20       24.086 ±      4.072    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                   100       true               true  thrpt   20        0.001 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm              100       true               true  thrpt   20       ≈ 10⁻³                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                     100       true               true  thrpt   20       50.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                      100       true               true  thrpt   20       47.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                              1000       true               true  thrpt   20  2163007.836 ± 133154.071   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                               1000       true               true  thrpt   20      157.132 ±      9.667  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                          1000       true               true  thrpt   20       80.004 ±      0.001    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                      1000       true               true  thrpt   20      157.292 ±     19.076  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                 1000       true               true  thrpt   20       80.193 ±      9.078    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                  1000       true               true  thrpt   20        0.001 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm             1000       true               true  thrpt   20        0.001 ±      0.001    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                    1000       true               true  thrpt   20       72.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                     1000       true               true  thrpt   20       62.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                              3000       true               true  thrpt   20  1047183.689 ±  66845.257   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                               3000       true               true  thrpt   20       76.073 ±      4.854  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                          3000       true               true  thrpt   20       80.005 ±      0.001    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                      3000       true               true  thrpt   20       74.277 ±     17.840  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                 3000       true               true  thrpt   20       77.846 ±     17.266    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                  3000       true               true  thrpt   20        0.001 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm             3000       true               true  thrpt   20        0.001 ±      0.002    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                    3000       true               true  thrpt   20       34.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                     3000       true               true  thrpt   20       33.000                   ms
```

Old encoder:
```
Benchmark                                                                              (contentLength)  (masking)  (pooledAllocator)   Mode  Cnt        Score        Error   Units
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                                 0       true               true  thrpt   20  6294845.923 ± 167741.242   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                  0       true               true  thrpt   20      137.188 ±      3.661  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                             0       true               true  thrpt   20       24.001 ±      0.001    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                         0       true               true  thrpt   20      137.617 ±     13.907  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                    0       true               true  thrpt   20       24.086 ±      2.408    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                     0       true               true  thrpt   20        0.001 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm                0       true               true  thrpt   20       ≈ 10⁻⁴                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                       0       true               true  thrpt   20       63.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                        0       true               true  thrpt   20       61.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                                 4       true               true  thrpt   20  6248817.496 ±  79365.118   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                  4       true               true  thrpt   20      136.133 ±      1.730  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                             4       true               true  thrpt   20       24.001 ±      0.001    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                         4       true               true  thrpt   20      135.358 ±     11.667  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                    4       true               true  thrpt   20       23.867 ±      2.045    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                     4       true               true  thrpt   20        0.001 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm                4       true               true  thrpt   20       ≈ 10⁻⁴                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                       4       true               true  thrpt   20       62.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                        4       true               true  thrpt   20       59.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                                 8       true               true  thrpt   20  6031607.505 ±  84317.103   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                  8       true               true  thrpt   20      131.405 ±      1.849  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                             8       true               true  thrpt   20       24.001 ±      0.001    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                         8       true               true  thrpt   20      133.186 ±      8.496  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                    8       true               true  thrpt   20       24.326 ±      1.474    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                     8       true               true  thrpt   20        0.001 ±      0.001  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm                8       true               true  thrpt   20       ≈ 10⁻⁴                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                       8       true               true  thrpt   20       61.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                        8       true               true  thrpt   20       58.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                                32       true               true  thrpt   20  5525281.410 ± 436145.997   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                 32       true               true  thrpt   20      120.420 ±      9.504  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                            32       true               true  thrpt   20       24.001 ±      0.001    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                        32       true               true  thrpt   20      122.333 ±     15.564  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                   32       true               true  thrpt   20       24.451 ±      2.778    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                    32       true               true  thrpt   20        0.002 ±      0.002  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm               32       true               true  thrpt   20       ≈ 10⁻⁴                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                      32       true               true  thrpt   20       56.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                       32       true               true  thrpt   20       54.000                   ms
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame                                               100       true               true  thrpt   20  3552029.630 ± 438693.405   ops/s
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate                                100       true               true  thrpt   20      258.051 ±     31.878  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.alloc.rate.norm                           100       true               true  thrpt   20       80.004 ±      0.001    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space                       100       true               true  thrpt   20      257.798 ±     36.735  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Eden_Space.norm                  100       true               true  thrpt   20       79.769 ±      4.887    B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space                   100       true               true  thrpt   20        0.001 ±      0.001  MB/sec
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.churn.G1_Survivor_Space.norm              100       true               true  thrpt   20       ≈ 10⁻⁴                 B/op
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.count                                     100       true               true  thrpt   20      118.000               counts
WebSocketFrame08EncoderBenchmark.writeWebSocketFrame:·gc.time                                      100       true               true  thrpt   20      105.000                   ms
…